% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sets.R
\name{make_riab_sets}
\alias{make_riab_sets}
\title{Create Riemann resamples}
\usage{
make_riab_sets(data, min_size, max_size, step, ...)
}
\arguments{
\item{data}{sf object}

\item{min_size}{numeric (min polygon radius)}

\item{max_size}{numeric (max polygon radius)}

\item{step}{numeric (interval for cellsize steps)}

\item{...}{passed to [sf::st_make_grid()]}
}
\description{
Create Riemann resamples
}
