% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_imports.R
\docType{package}
\name{riab-package}
\alias{riab}
\alias{riab-package}
\title{riab: Multi-Scale Assessments of Spatial Model Accuracy}
\description{
Functions aiding in multi-scale assessments of spatial model accuracy. By reporting spatially-explicit accuracy for predictions aggregated at multiple scales, users are able to provide more useful descriptions of how effectively their models represent the areas model consumers care about.
}
\keyword{internal}
